properly shutting down with docker now

+++ bin/communigate.sh
@@ -1,6 +1,24 @@
 #!/usr/bin/env sh
+APPLICATION="/opt/CommuniGate"
+BASEFOLDER="/var/CommuniGate"
+SUPPLPARAMS=
+[ -f ${APPLICATION}/CGServer ] || exit 0
+ulimit -c 2097151
+umask 0
 rm -f /var/CommuniGate/ProcessID
-/etc/init.d/CommuniGate start
-sleep 5
-trap "echo Caught Trap && /etc/init.d/CommuniGate stop" 1 2 3 6 15
-tail --pid=`cat /var/CommuniGate/ProcessID` -f /dev/null
+if [ -d ${BASEFOLDER} ] ; then
+    echo "${BASEFOLDER} already exists..."
+else
+    echo "Creating the CommuniGate Base Folder..."
+    mkdir ${BASEFOLDER}
+    chgrp mail ${BASEFOLDER}
+    chmod 2770 ${BASEFOLDER}
+fi
+echo "Starting CommuniGate Pro"
+exec ${APPLICATION}/CGServer \
+    --Base ${BASEFOLDER} \
+#   --Daemon ${SUPPLPARAMS} \
+#	--ClusterBackend
+#	--ClusterFrontend
+
+#    touch /var/lock/subsys/CommuniGate
